+++ lib/lerk/hive_interface.rb
@@ -113,10 +113,14 @@ EOF
         return
       end
 
-      steam_id ||= event.author.username
-      @logger.command(event, 'hive_query', { identifier: steam_id })
+      if discord_user.steam_account_id
+        steam_id ||= discord_user.steam_account_id
+      else
+        steam_id ||= event.author.username
+        @logger.command(event, 'hive_query', { identifier: steam_id })
+      end
 
-      account_id = resolve_account_id(steam_id)
+      account_id = Util.resolve_steam_account_id(steam_id)
       if account_id.nil?
         msg = "Could not convert #{ steam_id } to account ID, please try another."
         msg << steamid_help_message(event)
@@ -159,15 +163,6 @@ EOF
         end
       else
         PLAINTEXT_MESSAGE_TEMPLATE % args
-      end
-    end
-
-    def self.resolve_account_id(s)
-      begin
-        SteamID.from_string(s, api_key: Config::HiveInterface::STEAM_WEB_API_KEY).account_id
-      rescue WebApiError, ArgumentError => e
-        puts "Error: Could not convert #{ s } to account ID: #{ e.message }"
-        nil
       end
     end